requirements.txt
streamlit
anthropic
python-dotenv
PyPDF2
pypdf